requirement.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
astroid==2.11.5
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cssselect2==0.6.0
dill==0.3.5.1
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
django-social-share==2.2.1
django-taggit==3.0.0
future==0.18.2
html5lib==1.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.9.0
mccabe==0.7.0
oscrypto==1.3.0
pep8==1.7.1
Pillow==9.1.1
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pylint==2.14.1
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.10
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.0
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.7
